requirements.txt
gradio==5.24.0
pyyaml==6.0.1
anthropic==0.18.1
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
ollama==0.1.5